Updated Macro.c to read from mac.txt from a micro SD card
Macro_Init() reads the contents of mac.txt, which is contained in a micro SD card. The initializer also outputs the names of each macro in a 6x4 grid on the ST7735. Additionally ST7735.c has been updated to invert drawn characters, as the ST7735 in this application will be flipped 180 degrees.